// output/interfaces/wasi-filesystem-types.d.ts
export type Filesize = bigint;
// ...
export type ErrorCode =
  | "access"
  | "would-block"
  | "already"
  | "bad-descriptor"
  | "busy"
  | "deadlock"
  | "quota"
  | "exist"
  | "file-too-large"
  | "illegal-byte-sequence"
  | "in-progress"
  | "interrupted"
  | "invalid"
  | "io"
  | "is-directory"
  | "loop"
  | "too-many-links"
  | "message-size"
  | "name-too-long"
  | "no-device"
  | "no-entry"
  | "no-lock"
  | "insufficient-memory"
  | "insufficient-space"
  | "not-directory"
  | "not-empty"
  | "not-recoverable"
  | "unsupported"
  | "no-tty"
  | "no-such-device"
  | "overflow"
  | "not-permitted"
  | "pipe"
  | "read-only"
  | "invalid-seek"
  | "text-file-busy"
  | "cross-device";
import type { OutputStream } from "./wasi-io-streams.js";
export type { OutputStream };
/**
 * # Variants
 *
 * ## `"normal"`
 *
 * ## `"sequential"`
 *
 * ## `"random"`
 *
 * ## `"will-need"`
 *
 * ## `"dont-need"`
 *
 * ## `"no-reuse"`
 */
export type Advice =
  | "normal"
  | "sequential"
  | "random"
  | "will-need"
  | "dont-need"
  | "no-reuse";
export interface DescriptorFlags {
  read?: boolean;
  write?: boolean;
  fileIntegritySync?: boolean;
  dataIntegritySync?: boolean;
  requestedWriteSync?: boolean;
  mutateDirectory?: boolean;
}
/**
 * # Variants
 *
 * ## `"unknown"`
 *
 * ## `"block-device"`
 *
 * ## `"character-device"`
 *
 * ## `"directory"`
 *
 * ## `"fifo"`
 *
 * ## `"symbolic-link"`
 *
 * ## `"regular-file"`
 *
 * ## `"socket"`
 */
export type DescriptorType =
  | "unknown"
  | "block-device"
  | "character-device"
  | "directory"
  | "fifo"
  | "symbolic-link"
  | "regular-file"
  | "socket";
import type { Datetime } from "./wasi-clocks-wall-clock.js";
export type { Datetime };
export type NewTimestamp =
  | NewTimestampNoChange
  | NewTimestampNow
  | NewTimestampTimestamp;
export interface NewTimestampNoChange {
  tag: "no-change";
}
export interface NewTimestampNow {
  tag: "now";
}
export interface NewTimestampTimestamp {
  tag: "timestamp";
  val: Datetime;
}
export type LinkCount = bigint;
export interface DescriptorStat {
  type: DescriptorType;
  linkCount: LinkCount;
  size: Filesize;
  dataAccessTimestamp?: Datetime;
  dataModificationTimestamp?: Datetime;
  statusChangeTimestamp?: Datetime;
}
export interface PathFlags {
  symlinkFollow?: boolean;
}
export interface OpenFlags {
  create?: boolean;
  directory?: boolean;
  exclusive?: boolean;
  truncate?: boolean;
}